+++ p2panda-rs/src/schema/schema_name.rs
@@ -4,12 +4,14 @@ use std::fmt;
 use std::fmt::Display;
 use std::str::FromStr;
 
+use serde::{Deserialize, Serialize, Serializer};
+
 use crate::schema::error::SchemaNameError;
 use crate::schema::validate::validate_name;
 use crate::Validate;
 
 /// A human readable schema name, used in the construction of `SchemaId`.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct SchemaName(String);
 
 impl SchemaName {
@@ -34,6 +36,7 @@ impl Validate for SchemaName {
         if !validate_name(&self.0) {
             return Err(SchemaNameError::MalformedSchemaName);
         }
+
         Ok(())
     }
 }
@@ -49,6 +52,31 @@ impl FromStr for SchemaName {
 impl Display for SchemaName {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.0)
+    }
+}
+
+impl Serialize for SchemaName {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(&self.to_string())
+    }
+}
+
+impl<'de> Deserialize<'de> for SchemaName {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        // Deserialize into string
+        let name: String = Deserialize::deserialize(deserializer)?;
+
+        // Check format
+        let schema_name = SchemaName::new(&name)
+            .map_err(|err| serde::de::Error::custom(format!("invalid schema name, {}", err)))?;
+
+        Ok(schema_name)
     }
 }